# 10-10EZ | Upgrade performance for App container