--- a/src/applications/hca/containers/App.jsx
+++ b/src/applications/hca/containers/App.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import RoutedSavableApp from 'platform/forms/save-in-progress/RoutedSavableApp';
@@ -10,11 +10,15 @@ import { useLoa3UserData } from '../hooks/useLoa3UserData';
 import content from '../locales/en/content.json';
 import formConfig from '../config/form';
 
-const App = props => {
-  const { children, location } = props;
-  const { isLoadingFeatureFlags } = useSelector(selectFeatureToggles);
-  const { isLoadingProfile } = useSelector(selectAuthStatus);
-  const isAppLoading = isLoadingFeatureFlags || isLoadingProfile;
+const App = ({ children, location }) => {
+  const { isLoadingFeatureFlags, isLoadingProfile } = useSelector(state => ({
+    isLoadingFeatureFlags: selectFeatureToggles(state).isLoadingFeatureFlags,
+    isLoadingProfile: selectAuthStatus(state).isLoadingProfile,
+  }));
+  const isAppLoading = useMemo(
+    () => isLoadingFeatureFlags || isLoadingProfile,
+    [isLoadingFeatureFlags, isLoadingProfile],
+  );
 
   // Fetch appropriate data for LOA3 users
   useLoa3UserData();
@@ -42,11 +46,8 @@ const App = props => {
 };
 
 App.propTypes = {
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node,
-  ]),
-  location: PropTypes.object,
+  location: PropTypes.object.isRequired,
+  children: PropTypes.node,
 };
 
 export default App;