## Are you removing, renaming or moving a folder in this PR?
- [x] No, I'm not changing any folders (skip to TeamSites and delete the rest of this section)
- [ ] Yes, I'm removing, renaming or moving a folder

Did you change site-wide styles, platform utilities or other infrastructure?
- [x] No
- [ ] Yes, and I used the [proxy-rewrite steps](https://github.com/department-of-veterans-affairs/vets-website/tree/main/src/applications/proxy-rewrite#that-sounds-normal-so-whats-the-proxy-all-about) to test the injected header scenario

## Summary

With one of OCTOs key goals in 2025 being to improve overall performance on VA.gov, the 10-10 team is doing an audit our own products to ensure we are as performant as can be. This PR performs some basic upgrades to the global App container of the 10-10EZ.

Based on testing done utilizing the React Profiler API the following gains were made:

- Base render time improved from 2.2ms to 2.1ms
- Re-render time improved from 0.8ms to 0.6ms

## Related issue(s)

department-of-veterans-affairs/va.gov-team#103774

## Acceptance criteria

- Component is as performant as possible in all environments.

### Quality Assurance & Testing

- [x] No sensitive information (i.e. PII/credentials/internal URLs/etc.) is captured in logging, hardcoded, or specs
- [x] Linting warnings have been addressed
- [x] [Accessibility testing](https://depo-platform-documentation.scrollhelp.site/collaboration-cycle/prepare-for-an-accessibility-staging-review) has been performed

### Error Handling

- [x] Browser console contains no warnings or errors
- [x] Events are being sent to the appropriate logging solution